SaveState: Fix comparison warning

+++ pcsx2/SaveState.cpp
@@ -315,7 +315,7 @@ memLoadingState::memLoadingState(const VmStateBuffer& load_from)
 // Loading of state data from a memory buffer...
 void memLoadingState::FreezeMem( void* data, int size )
 {
-	if (m_idx + size > m_memory.size())
+	if (static_cast<u32>(m_idx + size) > m_memory.size())
 		m_error = true;
 
 	if (m_error)
